Code that allows 40 random passengers to be placed on a tour bus according to certain degrees of proximity.
Generate a random number and assign the passenger in that index of the series to seat number 1. Assign the passenger with the least distance to seat 2, using the distance matrix, to the passenger placed in seat 1. In seats 3 and 4, assign the passengers who are the least distance away from the passenger occupying the seat on their left among the remaining passengers. When moving to the second row, for the 5th seat, assign the passenger with the least sum of their distances to the passengers placed in the seats (1 and 2) just above (in front of) and on the upper diagonal. 5, 1, 2 and 3 for the 6th seat; 6, 2, 3 and 4 for the 7th seat; For the 8th seat, assign the passengers in the 7th, 3rd and 4th seats with the least sum of their distances. Appointments will be made taking into account the passengers other than the settled ones.